Moving SearchBar component
