+++ src/components/UI/SearchBar/SearchBar.tsx
@@ -65,3 +65,5 @@ export const SearchBar: React.SFC<SearchBarProps> = (props) => {
     </form>
   );
 };
+
+export default SearchBar;
